# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Swagger",
        "description": "API CRUD Students",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/issues": {
            "get": {
                "tags": [
                    "Issues"
                ],
                "summary": "Get list of issues",
                "description": "Return list of issues",
                "operationId": "11a7f564fd4f47eae59de57da89f5336",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Issue Id",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "A-101"
                        }
                    },
                    {
                        "name": "create_date",
                        "in": "query",
                        "description": "Issue Create Date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "2014-11-19"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/IssueResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Issues"
                ],
                "summary": "Create new issue",
                "description": "Create a new issue record",
                "operationId": "4f61806b1a4ad48c588682be6540ba88",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "title"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "A-101"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Network issue"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Can't ping IP address 127.0.0.1. Please fix this."
                                    },
                                    "create_date": {
                                        "type": "string",
                                        "example": "2014-11-19"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "example": "2014-11-19"
                                    },
                                    "priority": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Assigned"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Record created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IssueResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/issues/sample": {
            "post": {
                "tags": [
                    "Issues"
                ],
                "summary": "Create sample Issue",
                "description": "Create a new issue record",
                "operationId": "77f839482e672a714762a3e9f27a8851",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "title"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "A-101"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Network issue"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Can't ping IP address 127.0.0.1. Please fix this."
                                    },
                                    "create_date": {
                                        "type": "string",
                                        "example": "2014-11-19"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "example": "2014-11-19"
                                    },
                                    "priority": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Assigned"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Record created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IssueResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/issues/{id}/table": {
            "delete": {
                "tags": [
                    "Issues"
                ],
                "summary": "Delete issue table",
                "description": "Delete issue by ID",
                "operationId": "b06a3398ff6fde838db84e9d5c45d6e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "IssueResource": {
                "properties": {
                    "id": {
                        "description": "Id of the Issue",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description",
                        "type": "string"
                    },
                    "create_date": {
                        "description": "Create date",
                        "type": "string"
                    },
                    "last_update_date": {
                        "description": "Last update date",
                        "type": "string"
                    },
                    "due_date": {
                        "description": "Due date",
                        "type": "string"
                    },
                    "priority": {
                        "description": "Priority",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Status",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}